+++ Tests/Unit/Domain/Model/Event/SingleEventTest.php
@@ -548,4 +548,181 @@ final class SingleEventTest extends UnitTestCase
 
         self::assertSame($value, $this->subject->getOwnerUid());
     }
+
+    /**
+     * @test
+     */
+    public function getRegistrationStartInitiallyReturnsNull(): void
+    {
+        self::assertNull($this->subject->getRegistrationStart());
+    }
+
+    /**
+     * @test
+     */
+    public function setRegistrationStartSetsRegistrationStart(): void
+    {
+        $model = new \DateTime();
+        $this->subject->setRegistrationStart($model);
+
+        self::assertSame($model, $this->subject->getRegistrationStart());
+    }
+
+    /**
+     * @test
+     */
+    public function hasAdditionalTermsAndConditionsInitiallyReturnsFalse(): void
+    {
+        self::assertFalse($this->subject->hasAdditionalTermsAndConditions());
+    }
+
+    /**
+     * @test
+     */
+    public function setAdditionalTermsAndConditionsSetsAdditionalTermsAndConditions(): void
+    {
+        $this->subject->setAdditionalTermsAndConditions(true);
+
+        self::assertTrue($this->subject->hasAdditionalTermsAndConditions());
+    }
+
+    /**
+     * @test
+     */
+    public function isMultipleRegistrationPossibleInitiallyReturnsFalse(): void
+    {
+        self::assertFalse($this->subject->isMultipleRegistrationPossible());
+    }
+
+    /**
+     * @test
+     */
+    public function setMultipleRegistrationPossibleSetsMultipleRegistrationPossible(): void
+    {
+        $this->subject->setMultipleRegistrationPossible(true);
+
+        self::assertTrue($this->subject->isMultipleRegistrationPossible());
+    }
+
+    /**
+     * @test
+     */
+    public function getNumberOfOfflineRegistrationsInitiallyReturnsZero(): void
+    {
+        self::assertSame(0, $this->subject->getNumberOfOfflineRegistrations());
+    }
+
+    /**
+     * @test
+     */
+    public function setNumberOfOfflineRegistrationsSetsNumberOfOfflineRegistrations(): void
+    {
+        $value = 123456;
+        $this->subject->setNumberOfOfflineRegistrations($value);
+
+        self::assertSame($value, $this->subject->getNumberOfOfflineRegistrations());
+    }
+
+    /**
+     * @test
+     */
+    public function getStatusByDefaultReturnsPlanned(): void
+    {
+        self::assertSame(EventInterface::STATUS_PLANNED, $this->subject->getStatus());
+    }
+
+    /**
+     * @return array<string, array{0: EventInterface::STATUS_*}>
+     */
+    public function statusDataProvider(): array
+    {
+        return [
+            'planned' => [EventInterface::STATUS_PLANNED],
+            'confirmed' => [EventInterface::STATUS_CONFIRMED],
+            'canceled' => [EventInterface::STATUS_CANCELED],
+        ];
+    }
+
+    /**
+     * @test
+     * @param EventInterface::STATUS_* $status
+     * @dataProvider statusDataProvider
+     */
+    public function setStatusSetsStatus(int $status): void
+    {
+        $this->subject->setStatus($status);
+
+        self::assertSame($status, $this->subject->getStatus());
+    }
+
+    /**
+     * @return array<string, array{0: EventInterface::STATUS_*}>
+     */
+    public function nonCanceledStatusDataProvider(): array
+    {
+        return [
+            'planned' => [EventInterface::STATUS_PLANNED],
+            'confirmed' => [EventInterface::STATUS_CONFIRMED],
+        ];
+    }
+
+    /**
+     * @test
+     * @param EventInterface::STATUS_* $status
+     * @dataProvider nonCanceledStatusDataProvider
+     */
+    public function isCanceledForNonCanceledStatusReturnsFalse(int $status): void
+    {
+        $this->subject->setStatus($status);
+
+        self::assertFalse($this->subject->isCanceled());
+    }
+
+    /**
+     * @test
+     */
+    public function isCanceledForCanceledStatusReturnsTrue(): void
+    {
+        $this->subject->setStatus(EventInterface::STATUS_CANCELED);
+
+        self::assertTrue($this->subject->isCanceled());
+    }
+
+    /**
+     * @test
+     */
+    public function getSpecialPriceInitiallyReturnsZero(): void
+    {
+        self::assertSame(0.0, $this->subject->getSpecialPrice());
+    }
+
+    /**
+     * @test
+     * @dataProvider validPriceDataProvider
+     */
+    public function setSpecialPriceSetsSpecialPrice(float $price): void
+    {
+        $this->subject->setSpecialPrice($price);
+
+        self::assertSame($price, $this->subject->getSpecialPrice());
+    }
+
+    /**
+     * @test
+     */
+    public function getSpecialEarlyBirdPriceInitiallyReturnsZero(): void
+    {
+        self::assertSame(0.0, $this->subject->getSpecialEarlyBirdPrice());
+    }
+
+    /**
+     * @test
+     * @dataProvider validPriceDataProvider
+     */
+    public function setSpecialEarlyBirdPriceSetsSpecialEarlyBirdPrice(float $price): void
+    {
+        $this->subject->setSpecialEarlyBirdPrice($price);
+
+        self::assertSame($price, $this->subject->getSpecialEarlyBirdPrice());
+    }
 }